added new routes, optimized modal(update required)
