+++ src/css/Form.css
@@ -3,6 +3,14 @@
     flex-direction: column;
 }
 
+.form_content {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+/* .form_content h2 */
+
 .form input {
     /* margin: 5px;  */
     /* transition: 0.5s; */
@@ -21,7 +29,6 @@
     border-bottom: 3px solid blue;
 }
 
-button.login_button[type=submit]{
-    align-self: center;
+button.login_button[type=submit] {
     margin: 5px;
 }
