+++ src/css/Modal.css
@@ -28,7 +28,6 @@
     box-shadow: 0 5px 20px 0 rgba(0, 0, 0, 0.04);
     display: flex;
     flex-direction: column;
-
 }
 
 .modal_title {
@@ -43,7 +42,6 @@
 .modal_title p {
     margin: 0px;
 }
-
 
 .close_button {
     text-align: center;
@@ -64,4 +62,8 @@
     color: white;
     cursor: pointer;
     transition: background-color 0.15s;
+}
+
+.modal_content{
+    display: initial;
 }
