+++ src/pages/App.jsx
@@ -1,13 +1,19 @@
 import React from 'react'
+import { Routes, Route, Outlet, Link } from "react-router-dom";
+
 import { Layout } from './Layout'
 import "../css/App.css"
+import { Main } from './Main'
+import { Login, Register, Reset } from './Auth';
+
 
 export default function App() {
   return (
-    <Layout>
-      <h1 className='title'>Main page</h1>
-      <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Nostrum dolor veritatis molestiae et quidem aut? Repellat iure consectetur minima, rerum quaerat nemo esse laudantium, veniam repudiandae amet possimus nihil molestias. Lorem ipsum dolor sit amet consectetur adipisicing elit. Sunt ab omnis sed. Voluptatem, voluptatibus repellendus enim, eligendi omnis sint, sed officiis cum earum perferendis nisi doloremque hic veritatis fuga reprehenderit. Lorem ipsum dolor sit amet consectetur adipisicing elit. Officiis, natus necessitatibus ex nemo tenetur reiciendis sapiente tempora recusandae harum, iure consectetur eius in cumque qui magni, nesciunt voluptatum porro dolorem? Lorem ipsum dolor sit amet consectetur adipisicing elit. Laboriosam ullam porro eos quasi nostrum. Unde maxime architecto eum incidunt corporis illo obcaecati. Iusto consectetur quam deserunt sequi corrupti, quae maiores! Lorem ipsum dolor sit, amet consectetur adipisicing elit. Accusantium soluta atque id fuga iusto rem quidem vitae temporibus quos, quia debitis quo! Deserunt quibusdam dignissimos ipsa blanditiis cumque repellat voluptate. Lorem ipsum dolor sit amet consectetur adipisicing elit. Perferendis corporis dignissimos at eum aut adipisci suscipit repellendus facere iure, maiores unde libero incidunt natus doloribus provident numquam vel commodi. Ipsa!</p>
-
-    </Layout>
+    <Routes>
+      <Route path="/" element={<Main />} />
+      <Route path="/login" element={<Login />} />
+      <Route path="/register" element={<Register />} />
+      <Route path="/reset" element={<Reset />} />
+    </Routes>
   )
 }
