+++ webpack.config.js
@@ -3,8 +3,10 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
     entry: path.join(__dirname, "src", "index.js"),
+    devtool: 'source-map',
     output: {
-        path: path.resolve(__dirname, "dist")
+        path: path.resolve(__dirname, "dist"),
+        filename: 'main.bundle.js', 
     },
     module: {
         rules: [
@@ -28,5 +30,13 @@ module.exports = {
     mode: 'development',
     resolve: {
         extensions: ['', '.js', '.jsx', '.css'],
-    }
+    },
+    devServer: {
+        static: {
+            directory: path.join(__dirname, 'public'),
+        },
+        compress: true,
+        port: 3000,
+        historyApiFallback: true,
+    },
 }
